Move `parse_start_time_from` to correct module
